requirements.txt
groq
gradio
pillow
requests
python-dotenv